+++ amethyst_audio/src/lib.rs
@@ -1,3 +1,5 @@
+#![warn(missing_docs)]
+
 //! Loading and playing of audio files.
 extern crate amethyst_assets;
 extern crate amethyst_core;