+++ amethyst_gltf/src/lib.rs
@@ -1,3 +1,7 @@
+//! A crate for loading GLTF format scenes into Amethyst
+
+#![warn(missing_docs)]
+
 extern crate amethyst_animation as animation;
 extern crate amethyst_assets as assets;
 extern crate amethyst_core as core;
@@ -82,9 +86,12 @@ impl GltfPrefab {
     }
 }
 
+/// A GLTF node extent
 #[derive(Clone, Debug)]
 pub struct GltfNodeExtent {
+    /// The beginning of this extent
     pub start: Point3<f32>,
+    /// The end of this extent
     pub end: Point3<f32>,
 }
 
@@ -98,6 +105,7 @@ impl Default for GltfNodeExtent {
 }
 
 impl GltfNodeExtent {
+    /// Extends this to include the input range.
     pub fn extend_range(&mut self, other: &Range<[f32; 3]>) {
         for i in 0..3 {
             if other.start[i] < self.start[i] {
@@ -109,6 +117,7 @@ impl GltfNodeExtent {
         }
     }
 
+    /// Extends this to include the provided extent.
     pub fn extend(&mut self, other: &GltfNodeExtent) {
         for i in 0..3 {
             if other.start[i] < self.start[i] {
@@ -120,14 +129,17 @@ impl GltfNodeExtent {
         }
     }
 
+    /// Returns the centroid of this extent
     pub fn centroid(&self) -> Point3<f32> {
         (self.start + self.end.to_vec()) / 2.
     }
 
+    /// Returns the 3 dimensional distance between the start and end of this.
     pub fn distance(&self) -> Vector3<f32> {
         self.end - self.start
     }
 
+    /// Determines if this extent is valid.
     pub fn valid(&self) -> bool {
         for i in 0..3 {
             if self.start[i] > self.end[i] {