+++ amethyst_renderer/src/pass/debug_lines/mod.rs
@@ -1,3 +1,4 @@
+pub use self::interleaved::DebugLinesParams;
 pub use self::interleaved::DrawDebugLines;
 
 mod interleaved;