+++ amethyst_renderer/src/pass/shaders/geometry/debug_lines.glsl
@@ -24,27 +24,24 @@ layout (points) in;
 layout (triangle_strip, max_vertices = 4) out;
 
 uniform vec3 camera_position;
-
-// The line width is 1 line width unit, which is one 400th the unit size
-const float WIDTH = 1.0 / 400.0;
-const float HALF_WIDTH = WIDTH / 2.0;
+uniform float line_width;
 
 void EmitLine (int id) {
     vec3 cam_dir = normalize(vertex_in[id].position - camera_position);
     vec3 right = normalize(cross(cam_dir, vertex_in[id].normal));
 
-    vec3 width_vector = right * HALF_WIDTH;
+    vec3 width_vector = right * line_width * 0.5;
 
     vertex.color = vertex_in[id].color;
     vec3 pos = vertex_in[id].position - width_vector;
     gl_Position = proj * view * vec4(pos, 1.0);
     EmitVertex();
-    
+
     vertex.color = vertex_in[id].color;
     pos = vertex_in[id].position + width_vector;
     gl_Position = proj * view * vec4(pos, 1.0);
     EmitVertex();
-    
+
     vertex.color = vertex_in[id].color;
     pos = vertex_in[id].position + vertex_in[id].normal - width_vector;
     gl_Position = proj * view * vec4(pos, 1.0);