+++ examples/sprites_ordered/main.rs
@@ -14,8 +14,8 @@ mod sprite_sheet_loader;
 use amethyst::{
     assets::{AssetStorage, Loader},
     core::{
-        cgmath::{Matrix4, Ortho, Point3, Transform as CgTransform, Vector3},
-        transform::{GlobalTransform, Transform, TransformBundle},
+        cgmath::{Ortho, Point3, Transform as CgTransform, Vector3},
+        transform::{Transform, TransformBundle},
     },
     ecs::prelude::Entity,
     input::{get_key, is_close_requested, is_key_down},
@@ -206,15 +206,15 @@ impl Example {
             (dim.width(), dim.height())
         };
 
+        let mut camera_transform = Transform::default();
+        camera_transform.translation = Vector3::new(0., 0., self.camera_z);
+
         let camera = world
             .create_entity()
-            .with(GlobalTransform(Matrix4::from_translation(
-                Vector3::new(0.0, 0.0, self.camera_z).into(),
-            )))
+            .with(camera_transform)
             // Define the view that the camera can see. It makes sense to keep the `near` value as
             // 0.0, as this means it starts seeing anything that is 0 units in front of it. The
-            // `far` value is the distance the camera can see facing the origin. You can use
-            // `::std::f32::MAX` to indicate that it has practically unlimited depth vision.
+            // `far` value is the distance the camera can see facing the origin.
             .with(Camera::from(Projection::Orthographic(Ortho {
                 left: 0.0,
                 right: width,
@@ -303,9 +303,7 @@ impl Example {
                 // Render info of the default sprite
                 .with(sprite_render)
                 // Shift sprite to some part of the window
-                .with(sprite_transform)
-                // Used by the engine to compute and store the rendered position.
-                .with(GlobalTransform::default());
+                .with(sprite_transform);
 
             // The `Transparent` component indicates that the pixel color should blend instead of
             // replacing the existing drawn pixel.