+++ travis.sh
@@ -1,3 +1,4 @@
+RUSTFLAGS="-D warnings"
 MDBOOK_RELEASE="v0.2.1/mdbook-v0.2.1-x86_64-unknown-linux-gnu.tar.gz"
 
 echo "Build and test without profiler"